Fixes issue where binary assets delivered through `sandbox` may not be properly encoded
