+++ changelog.md
@@ -4,6 +4,19 @@
 Also see: [Architect changelog](https://github.com/architect/architect/blob/master/changelog.md)
 ---
 
+## [Coming soon] 2019-09-09
+
+### Changed
+
+- Internal change to normalize response shapes from `http.proxy.public` + `http.proxy.read`
+
+
+### Fixed
+
+- Fixes issue where binary assets delivered via `sandbox` / root may not be properly encoded
+
+---
+
 ## [3.3.8] 2019-09-03
 
 ### Fixed
